Move non-reusable step definitions into the private FeatureContext.
